backend/requirements.txt
fastapi==0.115.4
sqlmodel==0.0.22
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pyjwt==2.9.0
bcrypt==4.2.0
psycopg2-binary==2.9.9
email-validator==2.2.0
pydantic[email]==2.10.1
pytest==8.3.4
httpx==0.28.1
agents-mcp==0.1.0
openai==1.55.0
better-auth==0.0.1-beta.99